deleted categorical FCM and additional notebooks
